refactor gate classes
